+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Next Release
 ------------------
 
 * Added support for **Amazon Simple Queue Service (Amazon SQS)**
-* Updated CallableWaiter to allow for an array of context data to be passed to the callable
+* Updated `AbstractWaiter` to dispatch `waiter.before_attempt` and `waiter.before_wait` events
+* Updated `CallableWaiter` to allow for an array of context data to be passed to the callable
 * Fixed issue #29 so that the stat cache is cleared before performing multipart uploads
 * Fixed issue #38 so that Amazon CloudFront URLs are signed properly